Check for orphans before deleting containers
